refactor: extract cache utilities
